fix(bff): convention cleanup from framework audit
- Add ExecuteDeleteAsync bypass justification comments (interceptor-awareness convention)
- Extract BffSecurityHeadersMiddleware to its own file (one type per file)
- Consolidate duplicate MaskSessionId into BffSessionEndpoints (remove from BffLoginEndpoints)
- Remove duplicate MaskSessionId tests that tested the removed copy
